Refactor signature generation & matcer
