+++ plugin/callgraph/callgraph.go
@@ -1,122 +1,180 @@
 package callgraph
 
 import (
-	"context"
 	"fmt"
+	"slices"
 
 	"github.com/safedep/code/core"
 	"github.com/safedep/dry/log"
 	sitter "github.com/smacker/go-tree-sitter"
 )
 
-type CallgraphCallback func(*CallGraph) error
+const namespaceSeparator = "//"
 
-type callgraphPlugin struct {
-	// Callback function which is called with the callgraph
-	callgraphCallback CallgraphCallback
+// CallGraphNode represents a single node in the call graph
+type CallGraphNode struct {
+	Namespace string
+	CallsTo   []string
+	TreeNode  *sitter.Node
 }
 
-// Verify contract
-var _ core.TreePlugin = (*callgraphPlugin)(nil)
+type ContentDetails struct {
+	StartLine   uint32
+	EndLine     uint32
+	StartColumn uint32
+	EndColumn   uint32
+	Content     string
+}
 
-func NewCallGraphPlugin(callgraphCallback CallgraphCallback) *callgraphPlugin {
-	return &callgraphPlugin{
-		callgraphCallback: callgraphCallback,
+// GetContentDetails returns the content details of the node
+// If tree sitter node is nil, it returns false indicating that the content details are not available
+// else, it returns the content details and true
+func (gn *CallGraphNode) GetContentDetails(treeData *[]byte) (ContentDetails, bool) {
+	if gn.TreeNode == nil {
+		return ContentDetails{}, false
+	}
+	return ContentDetails{
+		StartLine:   gn.TreeNode.StartPoint().Row,
+		EndLine:     gn.TreeNode.EndPoint().Row,
+		StartColumn: gn.TreeNode.StartPoint().Column,
+		EndColumn:   gn.TreeNode.EndPoint().Column,
+		Content:     gn.TreeNode.Content(*treeData),
+	}, true
+}
+
+func newCallGraphNode(namespace string, treeNode *sitter.Node) *CallGraphNode {
+	return &CallGraphNode{
+		Namespace: namespace,
+		CallsTo:   []string{},
+		TreeNode:  treeNode,
 	}
 }
 
-func (p *callgraphPlugin) Name() string {
-	return "CallgraphPlugin"
+type CallGraph struct {
+	FileName          string
+	Nodes             map[string]*CallGraphNode
+	Tree              core.ParseTree
+	assignmentGraph   assignmentGraph
+	classConstructors map[string]bool
 }
 
-var supportedLanguages = []core.LanguageCode{core.LanguageCodePython}
-
-func (p *callgraphPlugin) SupportedLanguages() []core.LanguageCode {
-	return supportedLanguages
-}
-
-func (p *callgraphPlugin) AnalyzeTree(ctx context.Context, tree core.ParseTree) error {
-	lang, err := tree.Language()
+func newCallGraph(fileName string, importedIdentifiers map[string]parsedImport, tree core.ParseTree) (*CallGraph, error) {
+	language, err := tree.Language()
 	if err != nil {
-		return fmt.Errorf("failed to get language: %w", err)
+		return nil, fmt.Errorf("failed to get language from parse tree: %w", err)
 	}
 
-	file, err := tree.File()
-	if err != nil {
-		return fmt.Errorf("failed to get file: %w", err)
+	builtIns := getBuiltins(language)
+
+	cg := &CallGraph{
+		FileName:          fileName,
+		Nodes:             make(map[string]*CallGraphNode),
+		Tree:              tree,
+		assignmentGraph:   *newAssignmentGraph(),
+		classConstructors: make(map[string]bool),
 	}
 
-	log.Debugf("callgraph - Analyzing tree for language: %s, file: %s\n", lang.Meta().Code, file.Name())
-
-	cg, err := buildCallGraph(tree, lang, file.Name())
-
-	if err != nil {
-		return fmt.Errorf("failed to build call graph: %w", err)
+	for identifier, importedIdentifier := range importedIdentifiers {
+		cg.AddNode(importedIdentifier.Namespace, importedIdentifier.NamespaceTreeNode)
+		if identifier == importedIdentifier.Namespace {
+			cg.assignmentGraph.AddIdentifier(importedIdentifier.Namespace, importedIdentifier.NamespaceTreeNode)
+		} else {
+			cg.assignmentGraph.AddAssignment(identifier, importedIdentifier.IdentifierTreeNode, importedIdentifier.Namespace, importedIdentifier.NamespaceTreeNode)
+		}
 	}
 
-	return p.callgraphCallback(cg)
+	for _, namespace := range builtIns {
+		cg.assignmentGraph.AddIdentifier(namespace, nil) // @TODO - Can't create sitter node for keywords
+	}
+
+	return cg, nil
 }
 
-// buildCallGraph builds a call graph from the syntax tree
-func buildCallGraph(tree core.ParseTree, lang core.Language, filePath string) (*CallGraph, error) {
-	astRootNode := tree.Tree().RootNode()
-
-	treeData, err := tree.Data()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get tree data: %w", err)
+func (cg *CallGraph) AddNode(identifier string, treeNode *sitter.Node) {
+	if _, exists := cg.Nodes[identifier]; !exists {
+		cg.Nodes[identifier] = newCallGraphNode(identifier, treeNode)
 	}
-
-	imports, err := lang.Resolvers().ResolveImports(tree)
-	if err != nil {
-		return nil, fmt.Errorf("failed to resolve imports: %w", err)
-	}
-
-	// Required to map identifiers to imported modules as assignments
-	importedIdentifiers := parseImports(imports, lang)
-
-	// log.Debugf("Imported identifier => namespace:")
-	// for identifier, parsedImport := range importedIdentifiers {
-	// 	log.Debugf("  %s => %s\n", identifier, parsedImport.Namespace)
-	// }
-
-	callGraph, err := NewCallGraph(filePath, importedIdentifiers, tree)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create call graph: %w", err)
-	}
-
-	// Add root node to the call graph
-	callGraph.AddNode(filePath, astRootNode)
-
-	processChildren(astRootNode, *treeData, filePath, callGraph, processorMetadata{})
-
-	return callGraph, nil
 }
 
-func processNode(node *sitter.Node, treeData []byte, currentNamespace string, callGraph *CallGraph, metadata processorMetadata) processorResult {
-	if node == nil {
-		return newProcessorResult()
+// AddEdge adds an edge from one function to another
+func (cg *CallGraph) AddEdge(caller string, callerTreeNode *sitter.Node, callee string, calleeTreeNode *sitter.Node) {
+	cg.AddNode(caller, callerTreeNode)
+	cg.AddNode(callee, calleeTreeNode)
+	if !slices.Contains(cg.Nodes[caller].CallsTo, callee) {
+		cg.Nodes[caller].CallsTo = append(cg.Nodes[caller].CallsTo, callee)
 	}
-
-	nodeProcessor, exists := nodeProcessors[node.Type()]
-	if exists {
-		return nodeProcessor(node, treeData, currentNamespace, callGraph, metadata)
-	}
-
-	// log.Debugf("Can't process %s with namespace: %s => %s", node.Type(), currentNamespace, node.Content(treeData))
-	return emptyProcessor(node, treeData, currentNamespace, callGraph, metadata)
 }
 
-func processChildren(node *sitter.Node, treeData []byte, currentNamespace string, callGraph *CallGraph, metadata processorMetadata) processorResult {
-	if node == nil {
-		return newProcessorResult()
+func (cg *CallGraph) PrintCallGraph() {
+	fmt.Println("Call Graph:")
+	for caller, node := range cg.Nodes {
+		fmt.Printf("  %s (calls)=> %v\n", caller, node.CallsTo)
+	}
+	fmt.Println()
+}
+
+func (cg *CallGraph) PrintAssignmentGraph() {
+	fmt.Println("Assignment Graph:")
+	for assignmentNamespace, assignmentNode := range cg.assignmentGraph.Assignments {
+		fmt.Printf("  %s => %v\n", assignmentNamespace, assignmentNode.AssignedTo)
+	}
+	fmt.Println()
+}
+
+type DfsResultItem struct {
+	Namespace string
+	Node      *CallGraphNode
+	Caller    *CallGraphNode
+	Depth     int
+	Terminal  bool
+}
+
+func (cg *CallGraph) DFS() []DfsResultItem {
+	visited := make(map[string]bool)
+	var dfsResult []DfsResultItem
+	cg.dfsUtil(cg.FileName, nil, visited, &dfsResult, 0)
+	return dfsResult
+}
+
+func (cg *CallGraph) dfsUtil(namespace string, caller *CallGraphNode, visited map[string]bool, result *[]DfsResultItem, depth int) {
+	if visited[namespace] {
+		return
 	}
 
-	childrenResults := newProcessorResult()
+	callgraphNode, callgraphNodeExists := cg.Nodes[namespace]
 
-	for i := 0; i < int(node.ChildCount()); i++ {
-		result := processNode(node.Child(i), treeData, currentNamespace, callGraph, metadata)
-		childrenResults.addResults(result)
+	// Mark the current node as visited and add it to the result
+	visited[namespace] = true
+	*result = append(*result, DfsResultItem{
+		Namespace: namespace,
+		Node:      callgraphNode,
+		Caller:    caller,
+		Depth:     depth,
+		Terminal:  !callgraphNodeExists || len(callgraphNode.CallsTo) == 0,
+	})
+
+	assignmentGraphNode, assignmentNodeExists := cg.assignmentGraph.Assignments[namespace]
+	if assignmentNodeExists {
+		// Recursively visit all the nodes assigned to the current node
+		for _, assigned := range assignmentGraphNode.AssignedTo {
+			cg.dfsUtil(assigned, caller, visited, result, depth)
+		}
 	}
 
-	return childrenResults
+	// Recursively visit all the nodes called by the current node
+	// Any variable assignment would be ignored here, since it won't be in callgraph
+	if callgraphNodeExists {
+		for _, callee := range callgraphNode.CallsTo {
+			cg.dfsUtil(callee, callgraphNode, visited, result, depth+1)
+		}
+	}
+}
+
+func (cg *CallGraph) GetInstanceKeyword() (string, bool) {
+	language, err := cg.Tree.Language()
+	if err != nil {
+		log.Errorf("failed to get language from parse tree: %v", err)
+		return "", false
+	}
+	return resolveInstanceKeyword(language)
 }
